+++ src/main.cpp
@@ -1,17 +1,21 @@
 #include <Geode/Geode.hpp>
 #include <Geode/modify/GJGarageLayer.hpp>
 #include <Geode/modify/GJShopLayer.hpp>
-#include "layers/ShopListAlert.hpp"
+#include "ShopListAlert.hpp"
 
 using namespace geode::prelude;
 
 // GARAGE LAYER
-class $modify(ModLayer, GJGarageLayer){
-	bool init(){
-		if(!GJGarageLayer::init()) return false;
+class $modify(ModLayer, GJGarageLayer)
+{
+	bool init()
+	{
+		if (!GJGarageLayer::init())
+			return false;
 		auto sl_garageButton = Mod::get()->getSettingValue<bool>("garage-button");
 
-		if(sl_garageButton){
+		if (sl_garageButton)
+		{
 			NodeIDs::provideFor(this);
 
 			auto menu = this->getChildByID("shards-menu");
@@ -20,8 +24,7 @@ class $modify(ModLayer, GJGarageLayer){
 			auto button = CCMenuItemSpriteExtra::create(
 				spr,
 				this,
-				menu_selector(ModLayer::onShopListButton)
-			);
+				menu_selector(ModLayer::onShopListButton));
 			button->setID("Shopping-List-Button");
 
 			menu->addChild(button);
@@ -31,15 +34,19 @@ class $modify(ModLayer, GJGarageLayer){
 		return true;
 	}
 
-	void onShopListButton(CCObject *){
+	void onShopListButton(CCObject *)
+	{
 		ShopRewardsListAlert::create(ShopType(-1))->show();
 	}
 };
 
 // SHOP LAYER
-class $modify(ShopLayer, GJShopLayer){
-	bool init(ShopType p0){
-		if(!GJShopLayer::init(p0)) return false;
+class $modify(ShopLayer, GJShopLayer)
+{
+	bool init(ShopType p0)
+	{
+		if (!GJShopLayer::init(p0))
+			return false;
 		auto winSize = CCDirector::sharedDirector()->getWinSize();
 
 		NodeIDs::provideFor(this);
@@ -48,20 +55,21 @@ class $modify(ShopLayer, GJShopLayer){
 		menu->setPosition({30.0f, 30.0f});
 		menu->setID("Shopping-List-Menu");
 		menu->setZOrder(5);
-		
+
 		auto button = CCMenuItemSpriteExtra::create(
 			CCSprite::create("SL_ModInfo.png"_spr),
 			this,
-			menu_selector(ShopLayer::onInfoButton));
+			menu_selector(ShopLayer::onInfo));
 		button->setID("Shopping-List-Button");
-		
+
 		menu->addChild(button);
 		menu->updateLayout();
 		this->addChild(menu);
 		return true;
 	}
 
-	void onInfoButton(CCObject * sender){
+	void onInfo(CCObject *sender)
+	{
 		//  log::debug("Value {}", sender->getTag());
 
 		ShopRewardsListAlert::create(m_type)->show();
@@ -69,3 +77,13 @@ class $modify(ShopLayer, GJShopLayer){
 		//	ShopRewardsListAlert::create()->show();
 	}
 };
+
+$on_mod(Loaded)
+{
+	listenForAllSettingChangesV3([](std::shared_ptr<SettingV3> setting)
+								 {
+		if (auto shoppingList = static_cast<ShopRewardsListAlert*>(CCScene::get()->getChildByIDRecursive("shopping-list-popup")))
+		{
+			shoppingList->onNavButton(shoppingList->getChildByIDRecursive("navigation-menu")->getChildByTag(shoppingList->m_currentPage));
+		} });
+}